feat: ui panel work done
