--- src/sections/account/ApiTokenSection/ApiTokenSection.module.scss
@@ -1,15 +1,0 @@
-.exp-date {
-  display: flex;
-  gap: 1.5rem;
-  align-items: center;
-  padding-left: 0.5rem;
-  font-weight: bold;
-
-  time {
-    font-weight: normal;
-  }
-
-  @media (min-width: 768px) {
-    gap: 3rem;
-  }
-}
